customer model update

+++ src/main/java/com/olxseller/olx/DTO/ReviewDTO.java
@@ -15,17 +15,17 @@ public class ReviewDTO {
   private int id;
   @NotNull(message = "User ID can't be null")
   private int userId;
-  
+
   @NotNull(message = "Order ID can't be null")
   private int productId;
-  
+
   @NotNull(message = "rating can't be null")
-  @Min(value = 1,message = "Rating must be greater than 0")
-  private int rating;
-  
+  @Min(value = 1, message = "Rating must be greater than 0")
+  private float rating;
+
   @NotBlank(message = "User ID can't be null")
   private String review;
-  
+
   private LocalDateTime createdAt;
   private LocalDateTime updatedAt;
